Enh: Allow wrf_start/wrf_end to have nonzero minutes
Also added the ability to read .nc as well as .grib2 wrfnat files and removed model_dx. The latter was done because the map projection only works for dx = 1 km unless other keyword arguments (other than dx) are changed in the map projection function. This might make an interesting addition for later, but for now model_dx is removed to encourage users to not use wrfnat files with grid spacings other than 1 km.