Expand the max time window for interpolator and make tests more portable
Updated code greatly expands the UPP start and end times passed to
create_synthetic_obs.py to be +/- 24 hrs (previously was -3 to +1 hrs).
This ensures that CSV files can be created for synthetic obs with DHR
values in the window of -24 to +24.

Also refined the testing framework to be more portable. Previous version
always had to be run in the same directory on MSU HPC2.